# src/Data_Module/openapi.yaml
openapi: 3.1.0
info:
  title: Data Module API
  description: Intelligent Logistics - Data Module (Source of Truth)
  version: 1.0.0
servers:
  - url: http://10.255.32.82:8080/api/v1
    description: Production server
paths:
...
  /workers/login:
    post:
      tags: [Workers]
      summary: Login
      description: Worker login (operator or manager). Returns token.
      operationId: worker_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerLoginRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerLoginResponse'
  /workers/operators:
    get:
      tags: [Workers]
      summary: List Operators
      description: Lists operators.
      operationId: list_operators
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkerInfo'
...
  /workers/operators/{num_worker}:
    get:
      tags: [Workers]
      summary: Get Operator
      description: Gets information of a specific operator.
      operationId: get_operator
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
  /workers/operators/{num_worker}/current-shift/{gate_id}:
    get:
      tags: [Workers]
      summary: Get Operator Shift
      description: Gets operator's current shift for a gate.
      operationId: get_operator_shift
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
        - name: gate_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful Response
  /workers/operators/{num_worker}/shifts:
    get:
      tags: [Workers]
      summary: List Operator Shifts
      description: Lists shifts for an operator.
      operationId: list_operator_shifts
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
        - name: gate_id
          in: query
          schema:
            type: integer
            nullable: true
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful Response
  /workers/operators/{num_worker}/dashboard/{gate_id}:
    get:
      tags: [Workers]
      summary: Get Operator Dashboard
      description: Operator dashboard for a gate.
      operationId: get_operator_dashboard
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
        - name: gate_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorDashboard'
  /workers/managers:
    get:
      tags: [Workers]
      summary: List Managers
      description: Lists managers.
      operationId: list_managers
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkerInfo'
...
  /workers/managers/{num_worker}:
    get:
      tags: [Workers]
      summary: Get Manager
      description: Gets information of a specific manager.
      operationId: get_manager
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
  /workers/managers/{num_worker}/shifts:
    get:
      tags: [Workers]
      summary: List Manager Shifts
      description: Lists shifts supervised by a manager.
      operationId: list_manager_shifts
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Successful Response
  /workers/managers/{num_worker}/overview:
    get:
      tags: [Workers]
      summary: Get Manager Dashboard
      description: Manager dashboard/overview.
      operationId: get_manager_dashboard
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerOverview'
  /workers:
    get:
      tags: [Workers]
      summary: List All Workers
      description: Lists all workers (backoffice).
      operationId: list_all_workers
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: only_active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkerInfo'
    post:
      tags: [Workers]
      summary: Create New Worker
      description: Creates new worker (operator or manager).
      operationId: create_new_worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkerRequest'
      responses:
        '201':
          description: Created
  /workers/{num_worker}:
    get:
      tags: [Workers]
      summary: Get Worker
      description: Gets worker data.
      operationId: get_worker
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
    delete:
      tags: [Workers]
      summary: Deactivate Worker
      description: Deactivates a worker.
      operationId: deactivate_worker
      parameters:
        - name: num_worker
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
...
components:
  schemas:
...
    WorkerLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    WorkerLoginResponse:
      type: object
      required: [token, num_worker, name, email, active]
      properties:
        token:
          type: string
        num_worker:
          type: string
        name:
          type: string
        email:
          type: string
        active:
          type: boolean
    WorkerInfo:
      type: object
      required: [num_worker, name, email, role, active]
      properties:
        num_worker:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        active:
          type: boolean
    CreateWorkerRequest:
      type: object
      required: [num_worker, name, email, password, role]
      properties:
        num_worker:
          type: string
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        access_level:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
...
    OperatorDashboard:
      type: object
      required: [operator_num_worker, gate_id, date, upcoming_arrivals, stats]
      properties:
        operator_num_worker:
          type: string
        gate_id:
          type: integer
        date:
          type: string
        upcoming_arrivals:
          type: array
          items:
            type: object
        stats:
          type: object
          additionalProperties:
            type: integer
    ManagerOverview:
      type: object
      required: [manager_num_worker, date, active_gates, shifts_today, recent_alerts, statistics]
      properties:
        manager_num_worker:
          type: string
        date:
          type: string
        active_gates:
          type: integer
        shifts_today:
          type: integer
        recent_alerts:
          type: integer
        statistics:
          type: object
          additionalProperties:
            type: integer